refactor(core_logic): lib.rsをモジュール分割
## 変更内容
- CommandErrorをerror.rsに抽出
- list_images_in_folder, get_sibling_foldersをfs.rsに抽出
- lib.rsをpub mod + 再エクスポートのみに簡素化

## 技術的変更
- core_logic/src/error.rs: 新規作成（CommandError）
- core_logic/src/fs.rs: 新規作成（FS関数 + テスト）
- core_logic/src/lib.rs: pub mod + pub use のみ

## 検証結果
- cargo test: 全テストPASS（17 + core_logic単体4）
- 既存の外部参照は再エクスポートにより互換性維持
